Cambio int a double

+++ Ln0.c
@@ -17,7 +17,7 @@ double ln(double n)
 } 
 int main(int argc, char *argv[]){
 	int i;// para guardar el numero ingresado
-	int num; //para guardar la cantidad de iteraciones
+	double num; //para guardar la cantidad de iteraciones
 	int a; //para el ciclo for
 	int b; //para leer usando getopt
 	int errflg=0; //para gestionar los ingresos incorrectos
@@ -56,7 +56,7 @@ int main(int argc, char *argv[]){
 	  
 	  double total; 
 	  for (a=0; a < i; a++){ //todas las iteraciones (i=numero de iteraciones ingresadas)
-	    total = ln(num); 
+	      total = ln(num); 
 	  }
 	printf("%f\n", total); 
 	}
